requirements.txt
# Core dependencies
python-telegram-bot==21.9
openai==1.58.1
anthropic==0.40.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Security & Encryption
cryptography==42.0.0

# Calendar & Contacts providers
caldav==1.3.9
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
msal==1.26.0

# Excel handling
openpyxl==3.1.2
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
tenacity==8.3.0
tiktoken==0.7.0
PyYAML==6.0.1
structlog==24.1.0
pytz==2023.3

# vCard/vObject for contacts
vobject==0.9.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
